Indicate in error messages that account identifiers may be ambiguous
Users have complained that using a non-unique name for an account in the
REST API is treated as if such an account simply wouldn't exist. That
might be confusing for users.

For this reason, we now mention additionally that the identifier might
be ambiguous.

Ideally, we would distinguish between not finding an account at all and
an ambiguous identifier with two different tailored messages.
Implementing that proved to be more effort (and more complex code as a
result) than is worth for the benefit of a clear indication to users.

Change-Id: I3afaccf3cc70c443ece6c5cbb67a3b561e79b3a7

Former-commit-id: 50d34bffb2c17145d70eb19a05831fb27b65d58b