Provide `mergeable` to SHOW_CHANGE handlers
When the SKIP_MERGEABLE option was enabled for change detail requests,
the components in PG that depend on mergeability were refactored to get
mergeability info from properties besides the change detail. However,
some plugins depended on mergeability and used the detail response to
get it.

With this change, SHOW_CHANGE event details provide a new third `info`
parameter, which is an object containing a `mergeable` boolean. This
new info parameter becomes the preferred way for plugins to get
mergeability information.

The change detail object provided to SHOW_CHANGE handlers is given a
`mergeable` boolean getter for backwards compatibility with existing
plugins. The getter emits a deprecation warning and should be removed
after plugins have migrated.

Some logic in `_handleShowChange` is unrolled from the loop.

Bug: Issue 8231
Change-Id: Id2680abcd8baf62e64b6638be1c478da0490098f

Former-commit-id: 95a4d12570a42f5d2c04f0af512c3ff445b4cac9