EventFactory: Use AccountCache#maybeGet instead of AccountCache#get
AccountCache#get returns an empty AccountState to represent a missing
account. The empty AccountState has neither a full name, a preferred
email nor a user name. This means in case of a missing account we
created an empty AccountAttribute. An empty AccountAttribute is no
better than no AccountAttribute. Actually it seems that it was the
intention to return no AccountAttribute in case of a missing account
since null was returned if account state was null. Only account state
was never null since an empty account state was given when the account
was missing.

Change-Id: I558f5fe083e47d2b54a2987fa555b03ad01ce19a
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 85a8ddf4a031bf3fed933afbde9ccbe7c2e054e4