Use AccountCache#maybeGet instead of AccountCache#get when firing events
AccountCache#get returns an empty AccountState to represent a missing
account. The empty AccountState has neither a full name, a preferred
email nor a user name. This means in case of a missing account we
created an empty AccountAttribute. An empty AccountAttribute is no
better than no AccountAttribute. Actually it seems that it was the
intention to use no AccountAttribute in case of a missing account since
null is used if account state is null. Only by use of AccountCache#get
account state was never null.

Change-Id: Ie16c1a9c7521db19de7bf79ef4a9493e1c3c48b2
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: a09712aab68ffd8a7cc8b3e42b7038659daee5ac