Make sure status chip tooltip text gets updated
In the event where a previous chip had a tooltip and a new chip did not,
the text would not get removed properly. This change ensures that
tooltip text is cleared out, and also removes some unecessary work that
was done previously, as an empty string also removes the tooltip.

Bug: Issue 8148
Change-Id: I251183c334ad57c38d3d131b7ae996f5c4db4e93

Former-commit-id: 97de73882403beaeb3247812ad674de27e268798