setup_gjf.sh: Add support for multiple version
We are in process of upgrading to new google-java-format version. To
simplify upgrade process, allow to set up multiple different versions
of gjf distribution.

To install currently used version 1.3 no parameter is needed

  $ tools/setup-gjf.sh

The resulting script has now version suffix:
  tools/format/google-java-format-1.3

To update to a newer version, version parameter can be passed now:

  $ tools/setup-gjf.sh 1.5

Note, that because of SHA1 check only supported versions can be
currently installed: 1.3 and 1.5.

Change-Id: Id2464cd88291f7f7376df98ca2a396e2458d07a4

Former-commit-id: 48a581fe9b8bb1cf1b75ec9fdd2c672ba61f7d26