Link account chips to owner search rather than user dashboard
Before user dashboards were supported in PG, account links navigated to
an owner query. When user dashboard support was added, account links
started redirecting to the user's dashboard (or the owner query if the
user was logged out).

With this change, account links always link to the owner query, but the
user header that appears over the results shows a link to that user's
dashboard if the viewer is logged in.

Bug: Issue 8237
Change-Id: If0edfb886755c9ae0bb82d990169d04b96bcd57e

Former-commit-id: 6d4ea5877ddf907fbeb358ee510b8eb28439dff7