Simplify BatchUpdateTest and CommitsCollectionTest
Both of these tests use exactly the setup provided by
InMemoryTestEnvironment. Use that JUnit rule to avoid duplicate code and
improve readability of those classes.

The previous code in BatchUpdateTest and CommitsCollectionTest closed
the created repository after use. This shouldn't be necessary as the
Javadoc of InMemoryRepository states that closing the repository has no
impact on its memory/GC.

Change-Id: I64f62a3acb38b18030f00ebb41b662be7c602900

Former-commit-id: d4967d7b90ef08d74d1349671b77b64a973c1710