Remove circular dependency between diff_summary and diff caches
In order to migrate to Caffeine, the circular dependency between
diff_summary and diff caches needs to be removed. Guava is ok with this
circular dependency but Caffeine is not and goes into an infinite loop.

1.diff_summary.get(id=1234)
2.diff_summary_loader.get(id=1234)
3.diff.get(id=1234)
4.diff_loader.get(id=1234)
5.diff_summary.put(id=1234)->blocked because cache is waiting for step 2 to complete.

This circular dependency is not mandatory, values will be populated in
the diff_summary when explicitly requested.

Change-Id: I059741bd3f7229a8cf86ed13e5ce2fabfa54c6c8

Former-commit-id: 5fcf53d75238f18e7b555d6070b8ed15559969ec