Remove RefControl#isBlocked.
This method is used only tests and in isOwner().

In isOwner, use the stronger canPerform() method.

Blocking a user is not a observable part of the permission system, so
it should not be tested. To check that a user is blocked, one should
grant permissions in a child project, and observe that the child
permission is ignored.

Change-Id: I66fb450901eec2169883724c55445fab8e5f7bfa

Former-commit-id: ab3121d29c12d54f9826ac5f5667e2e6c28baa73