ChangeData: Expect that Account.Id.fromRefPart may return null
In practice this should never happen since before calling
Account.Id.fromRefPart we verified that it's a change edit ref. Hence
this is not a bug-fix but rather improving readability.

We cannot annotate Account.Id.fromRefPart with @Nullable since the
Account class is used in the GWT UI where @Nullable is not available.
However the follow-up change modifies this method to return an Optional.

Change-Id: Iae9d34e4ee225decd6ab4367827c1505d18acbdb
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: a7c80b0003fe5c70e74bdd9a87539655a220977a