CommitValidators: Always check for Change-Id in subject line
A commit with an empty commit message except for a Change-Id line in
the subject was only rejected if "Require Change-Id" was enabled.

The consequence of the commit being accepted is that the change gets
assigned a new Change-Id automatically by Gerrit, and it doesn't match
the one in the commit message. If the user then amends the local commit
with a full commit message, moving the Change-Id line from the subject
to the correct place in the footer, on pushing it will result in a new
change being created rather than a new patch set on the existing change.

Modify the logic so that it always checks for a Change-Id in the
subject line, not only when "Require Change-Id" is enabled.

Bug: Issue 8279
Change-Id: I416143f6ca04c9f2ed72c6174ff77b8ffba4256d

Former-commit-id: 3e2ef67a7f8b62da81d3e6efc005ebe407cbb8a6