CreateChange: Only insert Change-Id if there isn't already one
When creating a new change, the Change-Id is inserted using JGit's
ChangeIdUtil. This does not detect an existing Change-Id in the
subject line, resulting in the commit message having two Change-Id
lines. The latter Change-Id, automatically added, is the one that
is actually used.

Add an extra check to detect the Change-Id in the subject line and
not add another one.

Bug: Issue 8284
Change-Id: Ic3c5953a86e8287c1b2228f406fdc54c2b4d914a

Former-commit-id: c8c55e2131638509dcd8d5b8b31afea636de311a