Exclude NoteDb meta refs from TagSet
Similar to normal change refs, meta refs are not tag-able. Excluding
them from the list of interesting refs in TagSet scales down on the
amount of work we have to do later on when calculating tags.

The fewer references we have prior to calculating the tags, the less
memory we'll need when using BitSets to deal with refs flags.

Change-Id: Idc3308e3378bfede57d95cd65457129484385794

Former-commit-id: 545ce5b6e01e17df75e80db1861f366f581e100c