Suppress "Not found" when editing a new file
When entering edit mode and trying to create a new file, a server error
is shown. This occurs because restAPI._getFileInRevision throws an error
if the file doesn't yet exist.

With this change, an `opt_suppressErrors` array is added to
`_getFileInRevision`, `getChangeURLAndSend`, and `send` to allow
optional suppression of server errors. This new array is then used in
`getFileContent`.

This change also adds logic to gr-editor-view for handling a non-ok res.

Bug: Issue 8288
Change-Id: Ic5f954440583fe72b3521394eca4862925b3bbe7

Former-commit-id: f4d8357d8882f1f67cce2d0cd7d26b0ce27e4ee8