Account: Remove @Column annotations and make members private
This class is no longer a ReviewDb class and hence these annotations are
no longer needed.

We cannot remove the @Column annotation from Account.Id. This because we
store a proto for changes in the change index which includes an account
ID and protobuf needs this annotation for encoding/decoding.

Change-Id: Id3ad4a3454b1fc84527bd3a578b9f8db85883cf8
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: f3d5600d0cb7864fc8f31f571f677a7d9dafd94a