ReviewerRecommender: Prevent NPE when changeNotes is null
The changeNotes passed into suggestReviewers is possibly null, and
there is a check for this later in the method before attempting to
remove the change owner and existing reviewers from the list of
recommendations.

However the changeNotes was dereferenced to get the changeId to pass
into the plugin provided suggestReviewers method.

The suggestReviewers method annotates changeId as @Nullable, so we
can simply pass null there when changeNotes is null.

Change-Id: I4946b858dac05132cc65ebc6f4966a2fa0b23c11

Former-commit-id: d19d0d307ad19c092d18795796f88e660d9d1681