IndexChangeIT: Test index of change after its owner loses access
Add a test to conirm that the mergeable field is correctly set in
the change info on a change reindexed after its owner loses access.

Note that in the use case reported in the issue, all changes are
reindexed by a call to the "index start changes" ssh command. It is
not possible to reproduce this in the acceptance tests because the
online indexer is not available. Instead the change is explicitly
reindexed by the admin user.

The test added here fails when change I4df53914b is not applied.

Bug: Issue 8381
Change-Id: Ie4cd261de5315385703c1e11f1903b635f9862b5

Former-commit-id: e00e882605bad8ff5f03b335125f509f508febb1