Periodically reindex groups on slaves
Replication to slaves happens on Git level so that Gerrit is not aware
of incoming replication events. But slaves need an updated group index
to resolve memberships of users for ACL validation. To keep the group
index in slaves up-to-date a background job periodically scans the group
refs in the All-Users repository to reindex groups if they are stale.
The ref states of the group refs are cached so that on each run deleted
groups can be detected and reindexed. This means callers of slaves may
observe outdated group information until the next indexing happens. The
interval on which group indexing is done is configurable by setting
'index.scheduledIndexer.interval' in 'gerrit.config'. By default group
indexing is done every 5 minutes. This delay in group updates shouldn't
be a problem. Most slave setups already have a delay for group updates
to take effect by caching groups for a short period of time (configured
by 'maxAge' on the group related caches).

Detecting group deletions that were replicated while the slave was
offline is not supported. This means if group refs are deleted while the
slave is offline these groups are not removed from the group index when
the slave is started. However since group deletion is not supported this
should never happen and one can always do an offline reindex before
starting the slave.

This approach of keeping the group index up-to-date on slaves has the
advantage that it's backwards compatible and existing slave setups
(including running slaves in a cluster) don't need to be modified.
An alternative approach that made the replication events go through
Gerrit was implemented in changes I4bfd298cea, Ic8f8dd100b and
Ie8916b4a2. This alternative solution was abandoned in favor of the
solution implemented in this change.

Change-Id: Ibac7af13dea0be76f4708d2064373c212fae305a
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 0f0de8624d20d11707a3ec25f02b0b4b3b7448e5