Avoid unnecessary creation of GroupControl instance
The method GroupControl.Factory#controlFor(AccountGroup.Id) was only
used by GroupsCollection#parseId. When this call is inlined, it becomes
obvious that the creation of a GroupControl instance is unnecessary.

This change has the additional benefit of reducing the use of
AccountGroup.Id in our code base and to keep new code from using
GroupControl.Factory#controlFor(AccountGroup.Id) instead of the
recommended method GroupControl.Factory#controlFor(AccountGroup.UUID).

Change-Id: I9c3b69769a567192ec4582e11bd774c942e1b068

Former-commit-id: a59d8b7fe3766471aae726dbc0845c523f001c1a