NoteDbMigrator: Actually close ReviewDb after scanning projects
The code claims it was doing this, but it wasn't, unless --threads=1 and
all work is done in the main thread. Change it to explicitly open and
close the DB. This results in an extra open/close in the single-thread
case, but that's negligible compared to the cost of rebuilding.

Change-Id: Ic05b0118c2589ff1c8205b1140d11fe1f3d946a6

Former-commit-id: 6b8e53fd66bedc65337a8fa7bd0e6973f5b90735