Make it more clear when in "edit mode"
Previously, there was no indication that "edit mode" was loaded. This
change colors the header blue in edit-related views. It also removes the
yellow coloration of the file list header when an edit patch set is
loaded. This color may or may not be replaced with another color later,
based on feedback from users.

Change-Id: I42c546a1e5809c853f71f6aa6e7ad198213d45c0

Former-commit-id: 122f8982f0e3c3a4d4ee0201b7dd02fa5b3a51b6