Batch index executor: make use of the value passed for threads requested
Previously, the batch index executor was always configured using the values
provided in gerrit.config for key "batchThreads". This did not align
with the expectations when using "--threads" option from command line
in case of reindexing.

With this change, it takes into account the value passed in the threads
parameter, if no value is passed, it falls back to the value configured
in gerrit.config. As the constructor with "threads" parameter is only used
for reindex and DummyIndexModule for a slave, it is safe to make use of
the "threads" parameter to configure the executor.

Bug: Issue 8319
Change-Id: I6f26cc516e1879e710d89b7ab7d9d668d1e1a86f

Former-commit-id: 03ead35c034cd712cc398db3734c87cfe1a258aa