ScheduleConfig: Return empty Optional for invalid configuration
Returning an invalid configuration was only needed so that callers could
write a log for this. However this is unneeded since ScheduleConfig
already writes logs for missing and invalid configuration. Now duplicate
logs are avoided and there is only a single place in ScheduleConfig that
writes logs about a missing/invalid schedule configuration.

This also fixes the logs that were written by ScheduleConfig as they
didn't include the subsection name if a subsection was given.

With this change callers of ScheduleConfig can no longer differentiate
between missing and invalid configuration but they shouldn't have a need
for this. If a schedule config is required callers can now easily use
orElse or orElseThrow on the returned Optional.

Change-Id: I810309187d16bce092792fb4968bb7ade005ae7c
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: b1f614a808e695e29d5184a00325d84892ae1344