ChangeEmail: Don't log a warning for valid reviewer fields that are ignored
For each reviewer we store 2 values in the reviewer field, one value
with the format "<reviewer-type>,<reviewer-id>" and one value with the
format "<reviewer-type>,<reviewer-id>,<timestamp>". For parsing we are
only interested in the "<reviewer-type>,<reviewer-id>,<timestamp>" value
and the "<reviewer-type>,<reviewer-id>" value is ignored, but shouldn't
generate a warning in the logs.

Change-Id: Ic63f2216c42889f1073b19fc9c97017c0b35a7b7
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 18a7e3c19429d40cc9dc60971c502d5d5b6b97aa