Send notification emails when inbound emails are rejected
With this commit, explanatory messages are sent when inbound messages
are rejected.
Four error types are defined:
PARSING_ERROR, INACTIVE_ACCOUNT, UNKNOWN_ACCOUNT, and INTERNAL_EXCEPTION.

PARSING_ERROR (probably the most frequent one) occurs when the Gerrit
metadatas can't be parsed.

INACTIVE_ACCOUNT occurs when the user's account is Inactive.

UNKNOWN_ACCOUNT occurs when zero or more than one accounts are found for
the incoming email address. This might be caused by multiple sources
providing the same user.

INTERNAL_EXCEPTION is used for all the other exceptions that can't be
described properly to the end user. For now, it is only fired when two
Changes are found with the same Id. This _should_ be a rare exception,
but it might occur.

This change also introduces a new MailHeader enum, and removes the
old MetadataName one. This allows for a cleaner way to define both Mail
headers and Gerrit "internal" metadata names.

Feature: Issue 8210
Change-Id: I48a081f2ce1be391b9f3ff991760740d5ada3357

Former-commit-id: 2a10edcddc2cc6de404878de0b95f08c51d5165a