Fix reviewer suggestion for private changes
At the moment the reviewer suggestion only suggests users that can see
the change. This is bad for private changes since only existing
reviewers can see the change. Instead of checking the visibility of the
change, check that the destination branch is visible. This is what we
have done ealier when we had draft changes. This basically reverts
change I8c223849c9, but also adds a test and fixes the comment.

Change-Id: I0245a1391b5c4a4dadb07c1ab2b71ce116adf03c
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 3ed5c01cdde0af30649dcdb98a9b24081aef5ce6