Group members can't be added as reviewers if a matched username exists
Previously when trying to add a group as reviewers, gerrit would first
query the input as a username. The group would not be added if gerrit
found exactly one username that matched the input. This change modifies
the condition such that the input group name can be queried if the
username found by gerrit based on the input is not a perfect match.

Bug: Issue 8390
Change-Id: I2abfc59086986ede60a33dea5c09ae2a8d130367

Former-commit-id: 5bb8aea425e58ad705b15474b6b0902078d7fd2c