Do not show backdrop in reviewer confirmation overlay
If the user adds a group with a size greater than the configured
addreviewer.maxWithoutConfirmation limit, Gerrit will show a confirm
dialog (nested within the reply dialog) to prevent accidentally adding
large numbers of reviewers.

When it was first implemented in I0a61819e78, the dialog enabled the
overlay backdrop (in addition to the reply dialog backdrop), but that
backdrop was either not working initially, or it stopped appearing in
the intervening time.

Problematically, newer versions of iron-overlay-behavior will attempt to
honor the backdrop option in the nested dialog, but does so in such a
way that the overlay appears over the dialog, and the buttons are not
clickable.

With this change, the backdrop is disabled.

Bug: Issue 8353
Change-Id: Ia30613f4ea99edde6d5c9a1c2261e8b5e16d32b4

Former-commit-id: 6147bd16643fc027cde8b8c3d9aec0591925e01c